+++ inlang/source-code/message-lint-rules/camelCaseId/src/isCamelCaseId.ts
@@ -1,0 +1,3 @@
+export const isCamelCaseId = (id: string) => {
+	return /^[a-z]+([A-Z][a-z]+)*$/.test(id)
+}