+++ inlang/source-code/paraglide/paraglide-js-adapter-sveltekit/example/src/lib/ParaglideSvelteKit.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import { setLanguageTag, onSetLanguageTag } from '@inlang/paraglide-js/sveltekit';
+  import { setLanguageTag, onSetLanguageTag } from '../paraglide/runtime';
   import { getContext, setContext } from "svelte"
   import { page } from '$app/stores'