+++ inlang/source-code/website/src/pages/g/@uid/@id/index.page.server.tsx
@@ -3,19 +3,17 @@ import { convert } from "@inlang/markdown"
 import type { PageContext } from "#src/renderer/types.js"
 import type { PageProps } from "./index.page.jsx"
 import type { MarketplaceManifest } from "@inlang/marketplace-manifest"
-import { redirect } from "vite-plugin-ssr/abort"
+import { redirect } from "vike/abort"
 
 export async function onBeforeRender(pageContext: PageContext) {
 	const item = registry.find(
 		(item: any) => item.uniqueID === pageContext.routeParams.uid
 	) as MarketplaceManifest & { uniqueID: string }
 
-	if (
-		!item ||
-		item.id.replaceAll(".", "-").toLowerCase() !== pageContext.routeParams.id?.toLowerCase()
-	) {
-		console.error("Guide not found")
-		throw redirect("/m/404")
+	if (!item) throw redirect("/m/404")
+
+	if (item.id.replaceAll(".", "-").toLowerCase() !== pageContext.routeParams.id?.toLowerCase()) {
+		throw redirect(`/g/${item.uniqueID}/${item.id.replaceAll(".", "-").toLowerCase()}`)
 	}
 
 	const text = await (