+++ inlang/source-code/website/src/pages/index/custom_section/Guides.tsx
@@ -1,5 +1,5 @@
 import { For } from "solid-js"
-import * as m from "@inlang/paraglide-js/website/messages"
+import * as m from "#src/paraglide/messages.js"
 import Card from "#src/interface/components/Card.jsx"
 import { registry } from "@inlang/marketplace-registry"
 
@@ -13,7 +13,9 @@ const Guides = () => {
 	return (
 		<div class="w-full flex gap-4 mt-4 md:mt-10 mb-10 flex-col-reverse md:flex-row">
 			<div class="w-full">
-				<p class="pb-2 text-surface-500 text-sm">{m.home_guides_title()}</p>
+				<h2 class="pb-4 text-surface-900 font-semibold text-2xl leading-snug tracking-tight">
+					{m.home_guides_title()}
+				</h2>
 				<div class="grid lg:grid-cols-4 md:grid-cols-2 gap-4">
 					<For each={getGuides()}>
 						{(guide) => {