+++ inlang/source-code/plugins/m-function-matcher/CHANGELOG.md
@@ -1,1 +1,18 @@
 # @inlang/plugin-m-function-matcher
+
+## 0.4.0
+
+### Minor Changes
+
+- 8bb459652: update m function reference matcher
+
+## 0.3.0
+
+### Minor Changes
+
+- 39beea7dd: change return type of extractMessageOptions
+
+### Patch Changes
+
+- Updated dependencies [cafff8748]
+  - @inlang/sdk@0.17.0