+++ inlang/source-code/plugins/t-function-matcher/CHANGELOG.md
@@ -1,1 +1,12 @@
 # @inlang/plugin-t-function-matcher
+
+## 0.3.0
+
+### Minor Changes
+
+- 39beea7dd: change return type of extractMessageOptions
+
+### Patch Changes
+
+- Updated dependencies [cafff8748]
+  - @inlang/sdk@0.17.0